rack: layouts: setup retrieving layouts when loading fx
